+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\index.js":"1","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\App.js":"2","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Header.js":"3","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Layout.js":"4","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Footer.js":"5","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\PokemonCard.js":"6"},{"size":172,"mtime":1612084070642,"results":"7","hashOfConfig":"8"},{"size":4303,"mtime":1612091836086,"results":"9","hashOfConfig":"8"},{"size":404,"mtime":1611853609340,"results":"10","hashOfConfig":"8"},{"size":808,"mtime":1612089677362,"results":"11","hashOfConfig":"8"},{"size":311,"mtime":1611853058145,"results":"12","hashOfConfig":"8"},{"size":1744,"mtime":1612092623902,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1mqy5l9",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\index.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\App.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Header.js",[],["27","28"],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Layout.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Footer.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\PokemonCard.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\index.js":"1","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\App.js":"2","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Header.js":"3","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Layout.js":"4","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Footer.js":"5","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\PokemonCard.js":"6","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\routs\\Game\\index.js":"7","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\routs\\Home\\index.js":"8","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\MenuHeader.js":"9","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Menu.js":"10","C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Navbar.js":"11"},{"size":172,"mtime":1612084070642,"results":"12","hashOfConfig":"13"},{"size":498,"mtime":1612291386296,"results":"14","hashOfConfig":"13"},{"size":626,"mtime":1612276310463,"results":"15","hashOfConfig":"13"},{"size":834,"mtime":1612291675539,"results":"16","hashOfConfig":"13"},{"size":311,"mtime":1611853058145,"results":"17","hashOfConfig":"13"},{"size":1752,"mtime":1612279160975,"results":"18","hashOfConfig":"13"},{"size":231,"mtime":1612291528566,"results":"19","hashOfConfig":"13"},{"size":4747,"mtime":1612278270959,"results":"20","hashOfConfig":"13"},{"size":408,"mtime":1612291280496,"results":"21","hashOfConfig":"13"},{"size":921,"mtime":1612291288171,"results":"22","hashOfConfig":"13"},{"size":403,"mtime":1612291580104,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"1mqy5l9",{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"45","messages":"46","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\index.js",[],["47","48"],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\App.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Header.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Layout.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Footer.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\PokemonCard.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\routs\\Game\\index.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\routs\\Home\\index.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\MenuHeader.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Menu.js",[],"C:\\Users\\danya\\Desktop\\Marathon\\my-app\\src\\Components\\Navbar.js",["49"],{"ruleId":"50","replacedBy":"51"},{"ruleId":"52","replacedBy":"53"},{"ruleId":"54","severity":1,"message":"55","line":11,"column":5,"nodeType":"56","endLine":11,"endColumn":75},"no-native-reassign",["57"],"no-negated-in-lhs",["58"],"jsx-a11y/anchor-is-valid","The href attribute is required for an anchor to be keyboard accessible. Provide a valid, navigable address as the href value. If you cannot provide an href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
