+++ src/Components/Header.js
@@ -1,6 +1,11 @@
 import  s from "./Header.module.css";
 
-const Header = ({title,descr})=>{
+const Header = ({title,descr,onclickButton})=>{
+
+    const handleClick = ()=>{
+        console.log("<Header>");
+        onclickButton("game");
+    }
     return(
         <>
         <header className={s.root}>
@@ -8,6 +13,9 @@ const Header = ({title,descr})=>{
             <div className={s.container}>
                {title && <h1>{title}</h1> }
                 {descr && <p>{descr}</p>}
+                <button onClick={handleClick}>
+                    Start
+                </button>
              </div>
         </header>
         </>
