+++ src/Components/Menu.js
@@ -1,0 +1,35 @@
+import s from  "./Menu.module.css";
+
+const Menu = ({isActive})=>{
+    return(
+        <div className={`${s.menuContainer}  ${isActive ? s.active : s.deactive}`}>
+            <div className={s.overlay} />
+            <div className={s.menuItems}>
+                <ul>
+                <li>
+                    <a href="#welcome">
+                    HOME
+                    </a>
+                </li>
+                <li>
+                    <a href="#game">
+                    GAME
+                    </a>
+                </li>
+                <li>
+                    <a href="#about">
+                    ABOUT
+                    </a>
+                </li>
+                <li>
+                    <a href="#contact">
+                    CONTACT
+                    </a>
+                </li>
+                </ul>
+            </div>
+            </div>
+    )
+}
+
+export default Menu;
