+++ src/Components/PokemonCard.js
@@ -1,24 +1,26 @@
 import React from "react";
 import s from "./PokemonCard.module.css";
+import cn from "classnames"
+
 import cardBack from "../images/card-back-side.jpg"
 import {useState} from "react"
 const PokemonCard = ({name,id,type,img,top,bottom,left,right})=>{
     const[active,setActive] = useState(false);
 
     const clicker = ()=>{
-       setActive(true)
+       setActive(!active)
     }
     return(
         <div className={s.root} onClick={clicker}>
-      <div className={`${s.pokemonCard} ${active ? s.active : ""}`}>
+      <div className={cn(s.pokemonCard , {[s.active] : active})}>
         <div className={s.cardFront}>
-            <div className={`${s.wrap} ${s.front}`}>
-                <div className={`${s.pokemon} ${s[type]}`}>
+            <div className={cn(s.wrap, s.front)}>
+                <div className={cn(s.pokemon,s.type)}>
                     <div className={s.values}>
-                        <div className={`${s.count} ${s.top}`}>{top}</div>
-                        <div className={`${s.count} ${s.right}`}>{right}</div>
-                        <div className= {`${s.count} ${s.bottom}`}>{bottom}</div>
-                        <div className={`${s.count} ${s.left}`}>{left}</div>
+                        <div className={ cn(s.count,s.top)}>{top}</div>
+                        <div className={cn(s.count, s.right)}>{right}</div>
+                        <div className= {cn(s.count, s.bottom)}>{bottom}</div>
+                        <div className={cn(s.count, s.left)}>{left}</div>
                     </div>
                     <div className={s.imgContainer}>
                         <img src={img} alt={name} />
@@ -33,7 +35,7 @@ const PokemonCard = ({name,id,type,img,top,bottom,left,right})=>{
         </div>
 
         <div className={s.cardBack}>
-            <div className={`${s.wrap} ${s.back}`}>
+            <div className={cn(s.wrap, s.back)}>
                 <img src={cardBack} alt="Сard Backed" />
             </div>
         </div>
